Before:

```
trebuchetPart1() result: 55386, execution time: 45.125µs
trebuchetPart2() result: 54824, execution time: 9.765458ms
```

After:

```
trebuchetPart1() result: 55386, execution time: 42.25µs
trebuchetPart2() result: 54824, execution time: 831.584µs
```